# src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Rain
  description: Rain API's Documentation
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
...
security:
  - bearerAuth: []
paths:
...
  /pools/create-pool:
    post:
      summary: "Create a new pool"
      description: "Create a new pool with the specified question, options, and other data. If the pool is private, an access code will be generated automatically."
      operationId: "createPool"
      tags:
        - "Pools"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                question:
                  type: string
                  description: "The question for the pool (e.g., 'Will the price of ETH rise in the next week?')"
                  example: "Will the price of ETH rise in the next week?"
                questionImage:
                  type: string
                  description: "URL of an image related to the question"
                  example: "https://example.com/question-image.jpg"
                options:
                  type: array
                  items:
                    type: object
                    properties:
                      optionName:
                        type: string
                        description: "The name of the option"
                        example: "Yes"
                      optionImage:
                        type: string
                        description: "URL of the image representing the option"
                        example: "https://example.com/yes-option.jpg"
                  description: "The possible options for the pool (e.g., ['Yes', 'No'])"
                  example:
                    - optionName: "Yes"
                      optionImage: "https://example.com/yes-option.jpg"
                    - optionName: "No"
                      optionImage: "https://example.com/no-option.jpg"
                tags:
                  type: array
                  items:
                    type: string
                  description: "Tags related to the pool"
                  example: ["CRYPTO", "MOVIES", "SPORTS"]
                creator:
                  type: string
                  format: objectid
                  description: "User ID of the creator"
                  example: "60d5f3a5b0a7ac001c2d51a3"
                isPrivate:
                  type: boolean
                  description: "Flag indicating whether the pool is private or public"
                  example: true
                liquidityMax:
                  type: number
                  description: "Maximum liquidity for the pool"
                  example: 10000
                startDate:
                  type: string
                  format: date-time
                  description: "The start date and time for the pool"
                  example: "2024-12-10T12:00:00Z"
                poolTypeData:
                  type: array
                  description: "Data specific to public pools"
                  items:
                    type: object
                    properties:
                      minimumOracles:
                        type: number
                        description: "Minimum number of oracles required for the pool"
                        example: 3
                      maximumOracles:
                        type: number
                        description: "Maximum number of oracles allowed for the pool"
                        example: 5
                      rewardType:
                        type: string
                        enum: ["fixedAmount", "baseAmountPlusBonus"]
                        description: "Type of reward for the pool"
                        example: "fixedAmount"
                      rewardDetails:
                        type: object
                        properties:
                          fixedAmount:
                            type: object
                            properties:
                              rewardAmountPerOracle:
                                type: number
                                description: "The reward amount per oracle"
                                example: 100
                          baseAmountPlusBonus:
                            type: object
                            properties:
                              baseAmountPerOracle:
                                type: number
                                description: "Base reward amount per oracle"
                                example: 50
                              associatedDealSize:
                                type: number
                                description: "Associated deal size for the pool"
                                example: 1000
                              bonusPercentage:
                                type: number
                                description: "Bonus percentage for the reward"
                                example: 10
                      timeZone:
                        type: string
                        description: "Time zone for the pool"
                        example: "UTC"
                      duration:
                        type: number
                        description: "Duration of the pool in days"
                        example: 7
              required:
                - question
                - questionImage
                - options
                - creator
                - isPrivate
                - liquidityMax
                - startDate
                - poolTypeData
      responses:
        201:
          description: "Pool created successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 201
                  message:
                    type: string
                    example: "Pool created successfully."
                  data:
                    type: object
                    properties:
                      _id:
                        type: string
                        description: "Unique identifier of the pool"
                        example: "60d5f3a5b0a7ac001c2d51a3"
                      question:
                        type: string
                        example: "Will the price of ETH rise in the next week?"
                      questionImage:
                        type: string
                        example: "https://example.com/question-image.jpg"
                      options:
                        type: array
                        items:
                          type: object
                          properties:
                            optionName:
                              type: string
                              example: "Yes"
                            optionImage:
                              type: string
                              example: "https://example.com/yes-option.jpg"
                      tags:
                        type: array
                        items:
                          type: string
                        example: ["cryptocurrency", "ETH", "prediction"]
                      creator:
                        type: string
                        example: "60d5f3a5b0a7ac001c2d51a3"
                      isPrivate:
                        type: boolean
                        example: true
                      liquidityMax:
                        type: number
                        example: 10000
                      startDate:
                        type: string
                        format: date-time
                        example: "2024-12-10T12:00:00Z"
                      poolTypeData:
                        type: array
                        items:
                          type: object
                          properties:
                            minimumOracles:
                              type: number
                              example: 3
                            maximumOracles:
                              type: number
                              example: 5
                            rewardType:
                              type: string
                              example: "fixedAmount"
                            rewardDetails:
                              type: object
                              properties:
                                fixedAmount:
                                  type: object
                                  properties:
                                    rewardAmountPerOracle:
                                      type: number
                                      example: 100
                                baseAmountPlusBonus:
                                  type: object
                                  properties:
                                    baseAmountPerOracle:
                                      type: number
                                      example: 50
                                    associatedDealSize:
                                      type: number
                                      example: 1000
                                    bonusPercentage:
                                      type: number
                                      example: 10
                            timeZone:
                              type: string
                              example: "UTC"
                            duration:
                              type: number
                              example: 7
                      poolTrxHash:
                        type: string
                        description: "Transaction hash for the pool creation on the blockchain"
                        example: "0x8c7749a314bc8ee0dbfa3f5397f3475f63186efeb64e7f1cdbb63e1d9cdaf89f"
                      contractAddress:
                        type: string
                        description: "Smart contract address where the pool is deployed"
                        example: "0x2dB2b55feB45d8e4F4ecD0F51769497e4F7f45F7"
        400:
          description: "Bad request, missing required fields or invalid data"
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: "Bad request, some required fields are missing or invalid."
        500:
          description: "Internal Server Error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: "Error creating pool: internal server error."
  /pools/accessPool:
    post:
      summary: "Access a private pool using the access code"
      description: "Access a private pool by providing the pool ID and the correct access code."
      operationId: "accessPool"
      tags:
        - "Pools"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                poolId:
                  type: string
                  description: "The ID of the pool you want to access"
                  example: "60c72b2f9b1d8a001c8cbb7d"
                accessCode:
                  type: string
                  description: "The access code for a private pool"
                  example: "a1b2c3d4"
              required:
                - poolId
                - accessCode
      responses:
        200:
          description: "Access granted to the pool"
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "Access granted"
                  data:
                    type: object
                    properties:
                      poolId:
                        type: string
                        example: "60c72b2f9b1d8a001c8cbb7d"
                      question:
                        type: string
                        example: "What will be the weather like next week?"
                      isPrivate:
                        type: boolean
                        example: true
        403:
          description: "Forbidden. Invalid access code"
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 403
                  message:
                    type: string
                    example: "Invalid access code for private pool"
        404:
          description: "Pool not found"
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: "Pool not found"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...